refactor: clean up faculty evaluation and related type aliases
