+++ features/faculty-evaluation/components/faculty-evaluation-list-screen.tsx
@@ -1,23 +1,13 @@
 "use client";
 
 import { useEffect, useMemo, useState } from "react";
-import { ChevronDown, Search } from "lucide-react";
 
-import { Button } from "@/components/ui/button";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuRadioGroup,
-  DropdownMenuRadioItem,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
-import { Input } from "@/components/ui/input";
-import { ALL_PROGRAMS_VALUE } from "@/features/faculty-analytics/constants/filters";
-import { EvaluationViewToggle } from "@/features/faculty-evaluation/components/evaluation-view-toggle";
 import { FacultyGrid } from "@/features/faculty-evaluation/components/faculty-grid";
 import { FacultyList } from "@/features/faculty-evaluation/components/faculty-list";
 import { FacultyListState } from "@/features/faculty-evaluation/components/faculty-list-state";
+import { FacultyEvaluationToolbar } from "@/features/faculty-evaluation/components/faculty-evaluation-toolbar";
 import { useFacultyEvaluationListViewModel } from "@/features/faculty-evaluation/hooks/use-faculty-evaluation-list-view-model";
+import { getEvaluatorQuestionnaireTypes } from "@/features/faculty-evaluation/lib/questionnaire-types";
 import type {
   FacultyEvaluationRoleContext,
   FacultyEvaluationViewMode,
@@ -25,7 +15,6 @@ import type {
 import { useQuestionnaireTypes } from "@/features/questionnaires/hooks/use-questionnaire-types";
 
 const VIEW_STORAGE_KEY = "faculytics-faculty-evaluation-view";
-const STUDENT_ONLY_QUESTIONNAIRE_TYPE_CODE = "FACULTY_FEEDBACK";
 
 function getInitialViewMode(): FacultyEvaluationViewMode {
   if (typeof window === "undefined") {
@@ -40,10 +29,7 @@ export function FacultyEvaluationListScreen({ role }: { role: FacultyEvaluationR
   const [viewMode, setViewMode] = useState<FacultyEvaluationViewMode>(getInitialViewMode);
   const questionnaireTypesQuery = useQuestionnaireTypes();
   const evaluatorQuestionnaireTypes = useMemo(
-    () =>
-      (questionnaireTypesQuery.data ?? []).filter(
-        (type) => type.code !== STUDENT_ONLY_QUESTIONNAIRE_TYPE_CODE
-      ),
+    () => getEvaluatorQuestionnaireTypes(questionnaireTypesQuery.data),
     [questionnaireTypesQuery.data]
   );
   const {
@@ -109,85 +95,21 @@ export function FacultyEvaluationListScreen({ role }: { role: FacultyEvaluationR
         </div>
       </div>
 
-      <div className="mt-8 flex flex-col gap-3 lg:flex-row lg:items-center lg:justify-between">
-        <div className="relative w-full lg:max-w-sm">
-          <Search className="pointer-events-none absolute top-1/2 left-3 size-4 -translate-y-1/2 text-muted-foreground" />
-          <Input
-            value={searchValue}
-            onChange={(event) => setSearchValue(event.target.value)}
-            placeholder="Search faculty name..."
-            className="pl-9"
-            aria-label="Search faculty name"
-          />
-        </div>
-        <div className="flex w-full flex-col gap-3 sm:flex-row lg:w-auto">
-          <DropdownMenu>
-            <DropdownMenuTrigger asChild>
-              <Button
-                variant="outline"
-                className="w-full min-w-0 justify-between px-4 py-2.5 font-sans text-sm sm:w-56"
-                disabled={!selectedSemesterId || programs.length === 0}
-              >
-                <span className="truncate">{selectedProgramLabel}</span>
-                <ChevronDown className="size-4" />
-              </Button>
-            </DropdownMenuTrigger>
-            <DropdownMenuContent
-              align="end"
-              className="w-[var(--radix-dropdown-menu-trigger-width)] min-w-0"
-            >
-              <DropdownMenuRadioGroup
-                value={selectedProgramId ?? ALL_PROGRAMS_VALUE}
-                onValueChange={setSelectedProgramId}
-              >
-                {programs.map((program) => (
-                  <DropdownMenuRadioItem
-                    key={program.id ?? ALL_PROGRAMS_VALUE}
-                    value={program.id ?? ALL_PROGRAMS_VALUE}
-                    className="font-sans text-sm"
-                  >
-                    {program.label}
-                  </DropdownMenuRadioItem>
-                ))}
-              </DropdownMenuRadioGroup>
-            </DropdownMenuContent>
-          </DropdownMenu>
-          <DropdownMenu>
-            <DropdownMenuTrigger asChild>
-              <Button
-                variant="outline"
-                className="w-full min-w-0 justify-between px-4 py-2.5 font-sans text-sm sm:w-56"
-                disabled={semesters.length === 0}
-              >
-                <span className="truncate">{selectedSemesterLabel}</span>
-                <ChevronDown className="size-4" />
-              </Button>
-            </DropdownMenuTrigger>
-            <DropdownMenuContent
-              align="end"
-              className="w-[var(--radix-dropdown-menu-trigger-width)] min-w-0"
-            >
-              <DropdownMenuRadioGroup
-                value={selectedSemesterId ?? ""}
-                onValueChange={setSelectedSemesterId}
-              >
-                {semesters.map((semester) => (
-                  <DropdownMenuRadioItem
-                    key={semester.id}
-                    value={semester.id}
-                    className="font-sans text-sm"
-                  >
-                    {semester.label}
-                  </DropdownMenuRadioItem>
-                ))}
-              </DropdownMenuRadioGroup>
-            </DropdownMenuContent>
-          </DropdownMenu>
-          {facultyRows.length > 0 && selectedSemesterId ? (
-            <EvaluationViewToggle value={viewMode} onValueChange={setViewMode} />
-          ) : null}
-        </div>
-      </div>
+      <FacultyEvaluationToolbar
+        searchValue={searchValue}
+        onSearchChange={setSearchValue}
+        programs={programs}
+        selectedProgramId={selectedProgramId}
+        selectedProgramLabel={selectedProgramLabel}
+        onProgramChange={setSelectedProgramId}
+        semesters={semesters}
+        selectedSemesterId={selectedSemesterId}
+        selectedSemesterLabel={selectedSemesterLabel}
+        onSemesterChange={setSelectedSemesterId}
+        viewMode={viewMode}
+        onViewModeChange={setViewMode}
+        showViewToggle={facultyRows.length > 0 && Boolean(selectedSemesterId)}
+      />
 
       <div
         className={
