+++ features/faculty-evaluation/index.ts
@@ -1,6 +1,6 @@
 export { FacultyEvaluationListScreen } from "@/features/faculty-evaluation/components/faculty-evaluation-list-screen";
 export { FacultyEvaluationDetailScreen } from "@/features/faculty-evaluation/components/faculty-evaluation-detail-screen";
 export type {
+  FacultyEvaluationListItem,
   FacultyEvaluationRoleContext,
-  FacultyEvaluationRow,
 } from "@/features/faculty-evaluation/types";
